+++ src/Sword.java
@@ -1,7 +1,7 @@
 public class Sword extends Weapon {
 
 
-    public Sword(String name, int weight, int value) {
+    public Sword(String name, double weight, double value) {
         super(name, weight, value);
         setWeaponType("Sword");
     }
