+++ src/Wand.java
@@ -1,7 +1,7 @@
 public class Wand extends Weapon {
 
-    public Wand(String name, int weight, int value) {
-        super( name, weight, value);
+    public Wand(String name, double weight, double value) {
+        super(name, weight, value);
         setWeaponType("Wand");
     }
